[Spring 체스 - 3단계] 김김(김기문) 미션 제출합니다. (#306)
* refactor: 콜렉션 URL 복수형으로 변경

* refactor: Application에서 Controller 분리, ChessController의 @Controller와 @RestController 분리

* style: PieceDao의 쿼리 예약어 소문자 -> 대문자 변경

* style: GameDao의 쿼리 예약어 소문자 -> 대문자 변경

* style: 미사용 주석 삭제

* refactor: 미사용 클래스, 메소드 삭제

* refactor: dto 클래스 -body 접미사 삭제

* refactor: Controller에서 Domain에 대한 의존성 Service로 위임

* refactor: 새 게임 시작시 201 코드를 응답

* refactor: 기물을 이동시키고 저장할 때, 삭제-> 생성이 아닌 수정으로 처리

* feat: 예상치 못한 예외에 대한 처리 로직 생성

* refactor: 예외 객체가 HttpStatus 코드를 가지지 않도록 변경

* fix: ChessGameService.moveTest() 깨지는 문제 해결

- Service 메소드의 시그니쳐를 변경하고 반영하지 않아 테스트가 깨지는 문제가 있었습니다.

* refactor: application-test.properties 패키지 위치 이동, h2 Console 옵션 해제

* refactor: DB game, piece 테이블의 고유 값 칼럼 int -> bigint로 변경

* feat: 새 게임을 통해 게임 제목을 받을 수 있는 기능 추가, 다시하기 버튼 삭제

* feat: 게임이 제목을 가지도록 변경

* refactor: ResponseEntity의 body() 메소드가 아닌 ok() 함수 안에 dto 객체 주입

* refactor: Rank.asList() 미사용으로 인해 삭제

* refactor: MoveRequest 파라미터 없는 기본 생성자 오버로딩, setter 메소드 생성

* refactor: 기물 이동 요청 URI 변경(/piece -> /pieces)

* style: PieceDao savePieces() 메소드 공백 줄 추가로 가독성 재고

* refactor: EndWithoutGame 상태 미사용으로 인한 삭제

* refactor: GameDaoTest.deleteTest() 불필요 코드 삭제

* refactor: GameDaoTest ChessGameManager 중복 초기화 제거

* refactor: 테스트 코드 actual, expected 순서가 잘못된 코드 수정

* test: 게임 목록 조회, 게임 생성에 대한 테스트 코드 작성

* chore: 스키마 변경 사항 업로드

* refactor: 게임이 종료 되었을 때만 승패 여부를 나타내는 matchResult를 응답하도록 변경

* test: ChessRestController MockMvc를 이용해 테스트 코드 작성

* refactor: ChessGameService의 public method가 dto객체를 리턴하도록 수정

* refactor: Game/Piece Repository 및 구현체에 Override 어노테이션 명시, savePiece() 불필요로 인해 삭제

* refactor: 테스트코드 일부 인스턴스 변수의 접근 제어자 수정

* test: java-chess 레포지토리의 도메인 테스트 이식

* fix: 출발 좌표와 목적지 좌표가 같을 때 '/ by zero' 예외가 발생하는 버그 픽스

* refactor: 도메인에서 발생 가능한 예외 클래스들의 패키지 이동

* refactor: 비어있는 제목의 게임을 생성할 수 없도록 수정

* refactor: ChessControllerAdvice 리턴 코드 ResponseEntity의 정적 팩토리 메소드를 이용하도록 수정

* refactor: 게임 생성시 201 코드와 함께 Resource URI를 같이 리턴

* refactor: 미사용 테스트 정리, 각주 정리 및 정돈, todo로 남겨둔 테스트 작성

* refactor: ChessBoard.move()가 MoveRoute가 아닌 출발지, 목적지 Position을 받도록 수정 및 도메인 클래스들의 테스트 리팩토링

* fix: 게임 생성 요청 테스트 실패 문제 해결

* style: ChessRestControllerMockMvcTest DisplayName 일부 수정

* test: 빈 제목의 게임 생성 요청시 4xx 응답 요청 테스트 코드 작성

* fix: 나이트의 이동 거리 제한이 없는 버그 픽스

- 방향만 일정하다면(북북동, 북북서, 남남동, 남남서, 동동북, 동동남, 서서북, 서서남), 1번을 초과한 행마로만 갈 수 있는 행마를 할 수 있었던 문제를 해결했습니다.

* fix: Knight의 행마 검증 로직의 순서 배치로 인해 정확하지 않은 예외 메세지가 발생하는 문제 해결

* style: ChessRestController 코드 포맷팅

+++ src/main/java/chess/dto/NewGameDto.java
@@ -6,19 +6,19 @@ import chess.domain.piece.Color;
 import java.util.Map;
 
 public class NewGameDto {
-    private final int gameId;
+    private final long gameId;
     private final Map<String, PieceDto> chessBoard;
     private final Color currentTurnColor;
     private final Map<Color, Double> colorsScore;
 
-    public NewGameDto(int gameId, Map<String, PieceDto> chessBoard, Color currentTurnColor, Map<Color, Double> colorsScore) {
+    public NewGameDto(long gameId, Map<String, PieceDto> chessBoard, Color currentTurnColor, Map<Color, Double> colorsScore) {
         this.gameId = gameId;
         this.chessBoard = chessBoard;
         this.currentTurnColor = currentTurnColor;
         this.colorsScore = colorsScore;
     }
 
-    public static NewGameDto from(ChessGameManager chessGameManager, int gameId) {
+    public static NewGameDto from(ChessGameManager chessGameManager, long gameId) {
         return new NewGameDto(
                 gameId,
                 ChessBoardDto.from(chessGameManager.getBoard()).board(),
@@ -26,7 +26,7 @@ public class NewGameDto {
                 chessGameManager.getStatistics().getColorsScore());
     }
 
-    public int getGameId() {
+    public long getGameId() {
         return gameId;
     }
